pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "reflexor"
dynamic = ["version"]
description = "Reflexor"
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [{ name = "Reflexor contributors" }]
dependencies = [
  "fastapi>=0.110,<1",
  "pydantic>=2.0,<3",
  "pydantic-settings>=2.0,<3",
  "httpx>=0.27,<1",
  "PyYAML>=6,<7",
  "prometheus-client>=0.20,<1",
  "structlog>=24.1,<25",
  "typer>=0.12,<1",
  "uvicorn>=0.30,<1",
  "SQLAlchemy[asyncio]>=2.0,<3",
  "aiosqlite>=0.20,<1",
  "alembic>=1.13,<2",
]

[project.scripts]
reflexor = "reflexor.cli.main:app"

[project.optional-dependencies]
dotenv = ["python-dotenv>=1.0,<2"]
postgres = ["asyncpg>=0.29,<1"]
redis = ["redis>=5,<6"]
hiredis = ["hiredis>=2,<3"]
otel = [
  "opentelemetry-api>=1.36,<2",
  "opentelemetry-sdk>=1.36,<2",
  "opentelemetry-exporter-otlp-proto-http>=1.36,<2",
]
dev = [
  "coverage[toml]==7.13.4",
  "mypy==1.19.1",
  "pytest==9.0.2",
  "pytest-asyncio==1.3.0",
  "pytest-cov==7.0.0",
  "respx==0.22.0",
  "ruff==0.15.2",
]

[tool.hatch.version]
path = "src/reflexor/version.py"

[tool.hatch.build.targets.wheel]
packages = ["src/reflexor"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--cov-report=term-missing --cov-report=xml --cov-fail-under=91"

[tool.coverage.run]
branch = true
source = ["src"]
omit = ["tests/*"]

[tool.coverage.report]
show_missing = true
skip_covered = true
fail_under = 91

[tool.mypy]
files = ["src"]
python_version = "3.11"
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_optional = true
strict_optional = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
strict_equality = true
show_error_codes = true

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "ASYNC"]

[tool.ruff.lint.isort]
known-first-party = ["reflexor"]
